Unstable force-directed layout implemented
Here we model it as a physical system, using force, acceleration and
velocities. However, this diverges as there is no dampening in the
system. The next thing to try is to take inspiration from Eades method
to only move nodes a fraction in the direction of their force,
essentially skipping the momentum in the system, which should make it
stable.